requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
idna==2.10
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
mariadb==1.0.5
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
Pillow==8.0.1
psycopg2==2.8.6
pycairo==1.20.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.8.3
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
